requirements.txt
asgiref==3.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==3.2.16
django-crispy-forms==1.14.0
idna==3.4
Pillow==9.4.0
python-dotenv==0.21.1
pytz==2022.6
requests==2.31.0
sqlparse==0.4.3
stripe==7.6.0
typing-extensions==4.7.1
urllib3==2.0.7
